script: Move some inspector methods from ScriptBuf to Script
